+++ src/app/[locale]/list/_components/AasListTableRow.spec.tsx
@@ -20,7 +20,7 @@ jest.mock('next/navigation', () => ({
 }));
 jest.mock('./../../../../lib/services/list-service/aasListApiActions');
 jest.mock('next-auth', jest.fn());
-jest.mock('./../../../../lib/services/aas-repository-service/aasRepositorySearchActions', () => ({
+jest.mock('lib/services/aas-repository-service/aasRepositoryActions', () => ({
     getThumbnailFromShell: jest.fn(() => Promise.resolve({ success: true, result: { fileType: '', fileContent: '' } })),
 }));
 describe('AasListTableRow', () => {
