+++ src/app/[locale]/viewer/discovery/_components/DiscoveryListView.tsx
@@ -4,7 +4,7 @@ import { useSearchParams } from 'next/navigation';
 import AssetNotFound from 'components/basics/AssetNotFound';
 import { encodeBase64 } from 'lib/util/Base64Util';
 import ListHeader from 'components/basics/ListHeader';
-import { searchAasInAllRepositories } from 'lib/services/aas-repository-service/aasRepositorySearchActions';
+import { searchAasInAllRepositories } from 'lib/services/aas-repository-service/aasRepositoryActions';
 import { useTranslations } from 'next-intl';
 import { LocalizedError } from 'lib/util/LocalizedError';
 import { AasListEntry } from 'lib/types/AasListEntry';
