+++ src/lib/services/aas-repository-service/AasRepositoryService.spec.ts
@@ -1,4 +1,4 @@
-import { AasRepositorySearchService } from 'lib/services/aas-repository-service/AasRepositorySearchService';
+import { AasRepositoryService } from 'lib/services/aas-repository-service/AasRepositoryService';
 import { createTestAas, createTestInfrastructure } from 'test-utils/TestUtils';
 import { getInfrastructuresIncludingDefault } from 'lib/services/database/infrastructureDatabaseActions';
 import { encodeBase64 } from 'lib/util/Base64Util';
@@ -26,7 +26,7 @@ describe('AasRepositorySearchService', () => {
                 infrastructureName: 'Test Infrastructure',
             };
 
-            const aasRepositorySearchService = AasRepositorySearchService.createNull([mockSearchResult]);
+            const aasRepositorySearchService = AasRepositoryService.createNull([mockSearchResult]);
 
             const result = await aasRepositorySearchService.searchInAllAasRepositories(encodeBase64(aasId));
 
@@ -48,10 +48,7 @@ describe('AasRepositorySearchService', () => {
                 infrastructureName: 'Test Infrastructure',
             };
 
-            const aasRepositorySearchService = AasRepositorySearchService.createNull([
-                mockSearchResult1,
-                mockSearchResult2,
-            ]);
+            const aasRepositorySearchService = AasRepositoryService.createNull([mockSearchResult1, mockSearchResult2]);
 
             const result = await aasRepositorySearchService.searchInAllAasRepositories(encodeBase64(aasId));
 
@@ -80,7 +77,7 @@ describe('AasRepositorySearchService', () => {
                 infrastructureName: 'Test Infrastructure 1',
             };
 
-            const aasRepositorySearchService = AasRepositorySearchService.createNull([mockSearchResult]);
+            const aasRepositorySearchService = AasRepositoryService.createNull([mockSearchResult]);
 
             const result = await aasRepositorySearchService.searchAASInMultipleRepositories(
                 encodeBase64(aasId),
@@ -94,7 +91,7 @@ describe('AasRepositorySearchService', () => {
 
         it('returns not found if aas is not in any repository', async () => {
             const aasId = 'nonExistentAasId';
-            const aasRepositorySearchService = AasRepositorySearchService.createNull([]);
+            const aasRepositorySearchService = AasRepositoryService.createNull([]);
 
             const result = await aasRepositorySearchService.searchAASInMultipleRepositories(
                 encodeBase64(aasId),
@@ -109,7 +106,7 @@ describe('AasRepositorySearchService', () => {
 
         it('returns error if no repository url is configured', async () => {
             const aasId = 'testAasId';
-            const aasRepositorySearchService = AasRepositorySearchService.createNull([]);
+            const aasRepositorySearchService = AasRepositoryService.createNull([]);
 
             const result = await aasRepositorySearchService.searchInAllAasRepositories(encodeBase64(aasId));
