+++ src/lib/services/aas-repository-service/AasRepositoryService.ts
@@ -15,24 +15,22 @@ export type RepoSearchResult<T> = {
     infrastructureName?: string;
 };
 
-export class AasRepositorySearchService {
+export class AasRepositoryService {
     private constructor(
         protected readonly getAasRepositoryClient: (basePath: string) => IAssetAdministrationShellRepositoryApi,
         private readonly log: typeof logger = logger,
     ) {}
 
-    static create(log?: typeof logger): AasRepositorySearchService {
+    static create(log?: typeof logger): AasRepositoryService {
         const searcherLogger = log?.child({ Service: 'RepositorySearchService' });
-        return new AasRepositorySearchService(
+        return new AasRepositoryService(
             (baseUrl) => AssetAdministrationShellRepositoryApi.create(baseUrl, mnestixFetch()),
             searcherLogger,
         );
     }
 
-    static createNull(
-        shellsInRepositories: RepoSearchResult<AssetAdministrationShell>[] = [],
-    ): AasRepositorySearchService {
-        return new AasRepositorySearchService((baseUrl) =>
+    static createNull(shellsInRepositories: RepoSearchResult<AssetAdministrationShell>[] = []): AasRepositoryService {
+        return new AasRepositoryService((baseUrl) =>
             AssetAdministrationShellRepositoryApi.createNull(
                 baseUrl,
                 shellsInRepositories.filter((value) => value.location == baseUrl).map((value) => value.searchResult),
